+++ Utils/GamePath.cs
@@ -173,7 +173,7 @@ namespace Utils
                 Game.VFREVOBETA => "Virtua Fighter 5 R.E.V.O. Beta",
                 Game.VFREVO => "Virtua Fighter 5 R.E.V.O.",
                 Game.StrangerThanHeaven => "Stranger Than Heaven",
-                Game.YakuzaKiwami3 => "Yakuza Kiwami 3",
+                Game.yakuzakiwami3 => "Yakuza Kiwami 3",
                 _ => "<unknown>"
             };
         }